+++ CHANGELOG.md
@@ -11,8 +11,9 @@ Released on ??? (tbd)
 - Fixed `Query` dependence on `SearchParams` (#1024) @kzrnm
 - Fixed `Hash` and `Search` in `Url` (#1021) @kzrnm
 - Fixed behavior of `Url` with empty query (#1016) @kzrnm
-- Added option to prevent self-closing in `XhtmlMarkupFormatter` (#1031) @Aizeren
 - Added option to parse custom elements in `head` (#1035)
+- Added option to prevent self-closing in `XhtmlMarkupFormatter` (#1031) @Aizeren
+- Added option to keep original attribute names / casing (#897)
 
 # 0.16.1